Separating the Loader code.
