+++ src/Director/DirectorApplication.php
@@ -21,10 +21,10 @@ use Director\Model\Server;
 use Director\Model\Service;
 use Director\Service\AppService;
 
+use Director\Config\DirectorConfigLoader;
+
 use Symfony\Component\Config\FileLocator;
-use Symfony\Component\Yaml\Yaml;
 use Symfony\Component\Yaml\Dumper;
-use Symfony\Component\Config\Loader\FileLoader;
 use Symfony\Component\Config\Loader\LoaderResolver;
 use Symfony\Component\Config\Loader\DelegatingLoader;
 
@@ -76,37 +76,7 @@ class DirectorApplication extends BaseApplication
    */
   private function loadData() {
 
-    // Use $_SERVER['director_config_path'] if present.
-    $data_directories = array();
-    if (isset($GLOBALS['_SERVER']['director_config_path'])) {
-      $data_directories[] = $GLOBALS['_SERVER']['director_config_path'];
-    }
-    // Default to the config folder within this application.
-    // @TODO: Not sure how to properly deal with this yet.
-    $data_directories[] = __DIR__ . '/../../config';
-
-    // Attempt to locate data file
-    $locator = new FileLocator($data_directories);
-    try {
-      $this->configPath = dirname(realpath($locator->locate('director.yml')));
-    }
-    // If there's an exception, create a config directory.
-    catch (\InvalidArgumentException $e) {
-      // Copy default config template to default config directory.
-      $default_path = realpath(__DIR__ . '/../../config-default');
-      $cmd = "cp -r {$default_path} {$data_directories[0]}";
-      system($cmd);
-
-      // Save the configPath.  If it doesn't exist, we have a problem.
-      $this->configPath = realpath($data_directories[0]);
-      if (empty($this->configPath) || !file_exists($this->configPath . '/director.yml')) {
-        throw new \Exception("Unable to find or create a config folder!");
-      }
-    }
-
-    // YML Loader
-    $loaderResolver = new LoaderResolver(array(new DirectorConfigLoader($locator)));
-    $loader = new DelegatingLoader($loaderResolver);
+    $loader = $this->getLoader();
 
     // Load core director config.
     $this->config = $loader->load($this->configPath . '/director.yml');
@@ -176,26 +146,45 @@ class DirectorApplication extends BaseApplication
       $this->services[$name]:
       NULL;
   }
-}
 
+  /**
+   * Get a config loader object.
+   *
+   * @return \Symfony\Component\Config\Loader\DelegatingLoader
+   * @throws \Exception
+   */
+  private function getLoader() {
 
-/**
- * Class DevShopConfigLoader
- * @package DevShop
- */
-class DirectorConfigLoader extends FileLoader
-{
-  public function load($resource, $type = null)
-  {
-    $configValues = Yaml::parse(file_get_contents($resource));
-    return $configValues;
-  }
+    // Use $_SERVER['director_config_path'] if present.
+    $data_directories = array();
+    if (isset($GLOBALS['_SERVER']['director_config_path'])) {
+      $data_directories[] = $GLOBALS['_SERVER']['director_config_path'];
+    }
+    // Default to the config folder within this application.
+    // @TODO: Not sure how to properly deal with this yet.
+    $data_directories[] = __DIR__ . '/../../config';
 
-  public function supports($resource, $type = null)
-  {
-    return is_string($resource) && 'yml' === pathinfo(
-      $resource,
-      PATHINFO_EXTENSION
-    );
+    // Attempt to locate data file
+    $locator = new FileLocator($data_directories);
+    try {
+      $this->configPath = dirname(realpath($locator->locate('director.yml')));
+    }
+      // If there's an exception, create a config directory.
+    catch (\InvalidArgumentException $e) {
+      // Copy default config template to default config directory.
+      $default_path = realpath(__DIR__ . '/../../config-default');
+      $cmd = "cp -r {$default_path} {$data_directories[0]}";
+      system($cmd);
+
+      // Save the configPath.  If it doesn't exist, we have a problem.
+      $this->configPath = realpath($data_directories[0]);
+      if (empty($this->configPath) || !file_exists($this->configPath . '/director.yml')) {
+        throw new \Exception("Unable to find or create a config folder!");
+      }
+    }
+
+    // YML Loader
+    $loaderResolver = new LoaderResolver(array(new DirectorConfigLoader($locator)));
+    return new DelegatingLoader($loaderResolver);
   }
 }
